requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai==1.12.0
yt-dlp==2023.12.30
whisper==1.1.10
supabase==2.3.1
python-multipart==0.0.7
jinja2==3.1.3
httpx==0.26.0